refactor: change 'navigation' package name to 'navigation_drawer' and move it to 'components'
