+++ app/src/main/java/org/manifest/manifest24/MainActivity.kt
@@ -7,8 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import org.manifest.manifest24.ui.navigation.NavigationDrawer
-import org.manifest.manifest24.ui.pages.HomePage
+import org.manifest.manifest24.ui.components.navigation_drawer.NavigationDrawer
 import org.manifest.manifest24.ui.theme.Manifest24Theme
 
 class MainActivity : ComponentActivity() {
